move canvas logic to seperate function
